Introduce CpuInfo::getCpuPhysicalCoreCount().
Hyper-threading, when enabled, makes the number of processors two
times the number of physical cores.

Member function getCpuPhysicalCoreCount() is introduced to get the
number of physical cores and it is used in "Dashboard" but not in
"Resources".

The way it works is that it counts the number of unique physical
id, core id pair in /proc/cpuinfo.

+++ stacer/Managers/info_manager.h
@@ -15,7 +15,7 @@ class InfoManager
 public:
     static InfoManager *ins();
 
-    quint8 getCpuCoreCount() const;
+    int getCpuCoreCount() const;
     QList<int> getCpuPercents() const;
     QList<double> getCpuLoadAvgs() const;
     double getCpuClock() const;
